inventory_scaffold3/requirements_additions.txt
pandas
openpyxl
Pillow
django-htmx
psycopg2-binary
